pyproject.toml
[tool.poetry]
name = "shameleon"
version = "0.2.0"
description = ""
authors = [
    "Jychp <jeremy.chapeau@resilience.care>"
]
readme = "README.md"
homepage = "https://github.com/jychp/shameleon"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
cryptography = "^42.0.5"
aiohttp = "^3.9.4"

[tool.poetry.group.dummyserver.dependencies]
fastapi = "^0.110.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
